laughter-detector/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.3.0
tensorflow==2.15.0
tensorflow-hub==0.15.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
cryptography==41.0.8
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
bandit==1.7.5
black==23.11.0
ruff==0.1.6
mypy==1.7.1
